[Fission] Pod Reaper dependency clean up and additional logging.
[git-p4: depot-paths = "//depot/service/java/projects/DistributedFileHandling/ContentProcessing/handler/podreaper/": change = 536761]

+++ main/impl/src/main/java/com/theplatform/dfh/cp/handler/reaper/impl/processor/ReaperProcessor.java
@@ -53,6 +53,11 @@ public class ReaperProcessor implements HandlerProcessor
     public void execute()
     {
         List<Pod> podsToDelete = podLookupFilter.performLookup();
+        if(podsToDelete == null || podsToDelete.size() == 0)
+        {
+            logger.info("No pods found for reaping.");
+            return;
+        }
         // TODO: break this up into smaller sets?
         podsToDelete.forEach(x -> logger.info("Attempting to delete pod: {}", x.getMetadata().getName()));
         kubernetesPodFacade.deletePods(podsToDelete);
